% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_helpers.R
\name{.desurv_make_hyper_grid}
\alias{.desurv_make_hyper_grid}
\title{Build and validate DeSurv hyperparameter grid}
\usage{
.desurv_make_hyper_grid(
  k_grid,
  lambda_grid,
  nu_grid,
  lambdaW_grid,
  lambdaH_grid
)
}
\description{
Build and validate DeSurv hyperparameter grid
}
\keyword{internal}
