% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_filter.R
\name{gene_filter}
\alias{gene_filter}
\title{Select highly expressed and variable genes}
\usage{
gene_filter(X, ngene = 1000)
}
\arguments{
\item{X}{Numeric matrix with genes as rows and samples as columns.}

\item{ngene}{Positive integer indicating how many genes to retain.}
}
\value{
A numeric matrix containing the top \code{ngene} genes ranked by
average expression and variability (ties broken by combined rank).
}
\description{
Select highly expressed and variable genes
}
