% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Multi-start initialization for DeSurv}
\usage{
init(
  X,
  y,
  delta,
  k,
  alpha,
  lambda,
  nu,
  lambdaW,
  lambdaH,
  seed = 123,
  tol_init = 1e-04,
  imaxit = 30,
  verbose = TRUE,
  ninit = 20,
  parallel = FALSE,
  ncores = NULL
)
}
\description{
Run \code{ninit} short DeSurv optimizations with at most \code{imaxit}
iterations each, using a relaxed tolerance \code{tol_init}, evaluate
C-index of each, and return the best initialization (W, H, beta).
Can optionally run the short initializations in parallel.
}
\keyword{internal}
