% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeSurv-package.R
\docType{package}
\name{DeSurv-package}
\alias{DeSurv}
\alias{DeSurv-package}
\title{DeSurv: Implements survival-driven deconvolution (NMF+cox model)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Amber Young \email{ayoung31@ad.unc.edu}

}
\keyword{internal}
